requirements.txt
selenium==4.22.0
webdriver-manager==4.0.2
pytest==8.2.2
pytest-xdist==3.6.1
allure-pytest==2.13.5
python-dotenv==1.0.1
pandas==2.2.2
openpyxl==3.1.2